test(core): extract FakeCrashReporter helper and cover Dio badResponse edge cases
