+++ test/core/errors/error_handler_test.dart
@@ -1,5 +1,3 @@
-import 'dart:developer';
-
 import 'package:flutter_test/flutter_test.dart';
 import 'package:nuntium/config/dependency_injection.dart';
 import 'package:nuntium/core/errors/crash_reporter.dart';
@@ -7,17 +5,7 @@ import 'package:nuntium/core/errors/error_handler.dart';
 import 'package:nuntium/core/errors/exceptions.dart';
 import 'package:nuntium/core/errors/failures.dart';
 
-class FakeCrashReporter implements CrashReporter {
-  @override
-  Future<void> reportError({
-    required dynamic exception,
-    required StackTrace? stackTrace,
-    dynamic reason,
-    bool fatal = false,
-  }) async {
-    log(exception.toString());
-  }
-}
+import '../../helpers/fake_crash_reporter.dart';
 
 void main() {
   setUpAll(() {
